requirements.txt
polars==0.17.15
apache-airflow[amazon]